vasi-sync: create a tests sync module
