+++ src/lib/vasi-sync/tests/scmutex-tests.rs
@@ -5,44 +5,7 @@
 //! [loom]: <https://docs.rs/loom/latest/loom/>
 use vasi_sync::scmutex::{SelfContainedMutex, SelfContainedMutexGuard};
 
-mod sync {
-    #[cfg(loom)]
-    pub fn model<F>(f: F)
-    where
-        F: Fn() + Sync + Send + 'static,
-    {
-        loom::model(move || {
-            f();
-            vasi_sync::sync::loom_reset();
-        });
-    }
-    #[cfg(not(loom))]
-    pub fn model<F>(f: F)
-    where
-        F: Fn() + Sync + Send + 'static,
-    {
-        f()
-    }
-
-    #[cfg(loom)]
-    pub use loom::sync::Arc;
-    #[cfg(not(loom))]
-    pub use std::sync::Arc;
-
-    #[cfg(loom)]
-    pub use loom::thread;
-    #[cfg(not(loom))]
-    pub use std::thread;
-
-    #[cfg(loom)]
-    pub fn rand_sleep() {}
-    #[cfg(not(loom))]
-    pub fn rand_sleep() {
-        std::thread::sleep(std::time::Duration::from_nanos(
-            rand::random::<u64>() % 10_000_000,
-        ));
-    }
-}
+mod sync;
 
 mod tests {
     use super::*;
